In GzipHelperTask and TarHelperTask do almost same lines in addElements() This class is for use template method pattern for only them, not zipHelperTask, RarHelperTask if there will be big differences, this change have to be roll-back.
