+++ app/src/main/java/com/amaze/filemanager/asynchronous/asynctasks/compress/TarHelperTask.java
@@ -1,5 +1,7 @@
 package com.amaze.filemanager.asynchronous.asynctasks.compress;
 
+import android.support.annotation.NonNull;
+
 import com.amaze.filemanager.adapters.data.CompressedObjectParcelable;
 import com.amaze.filemanager.filesystem.compressed.CompressedHelper;
 import com.amaze.filemanager.utils.OnAsyncTaskFinished;
@@ -19,7 +21,7 @@ import static com.amaze.filemanager.filesystem.compressed.CompressedHelper.SEPAR
  *         on 2/12/2017, at 00:40.
  */
 
-public class TarHelperTask extends CompressedHelperTask {
+public class TarHelperTask extends GzTarHelperTask {
 
     private String filePath, relativePath;
 
@@ -30,30 +32,10 @@ public class TarHelperTask extends CompressedHelperTask {
         this.relativePath = relativePath;
     }
 
+    @NonNull
     @Override
-    void addElements(ArrayList<CompressedObjectParcelable> elements) {
-        TarArchiveInputStream tarInputStream = null;
-        try {
-            tarInputStream = new TarArchiveInputStream(new FileInputStream(filePath));
-
-            TarArchiveEntry entry;
-            while ((entry = tarInputStream.getNextTarEntry()) != null) {
-                String name = entry.getName();
-                if (name.endsWith(SEPARATOR)) name = name.substring(0, name.length() - 1);
-
-                boolean isInBaseDir = relativePath.equals("") && !name.contains(SEPARATOR);
-                boolean isInRelativeDir = name.contains(SEPARATOR)
-                        && name.substring(0, name.lastIndexOf(SEPARATOR)).equals(relativePath);
-
-                if (isInBaseDir || isInRelativeDir) {
-                    elements.add(new CompressedObjectParcelable(entry.getName(),
-                            entry.getLastModifiedDate().getTime(), entry.getSize(), entry.isDirectory()));
-                }
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
+    TarArchiveInputStream getTarArchiveInputStream() throws IOException {
+        return new TarArchiveInputStream(new FileInputStream(filePath));
     }
 
 }
